requirements.txt
fastapi
uvicorn
python-pptx
groq
python-dotenv
requests
duckduckgo_search
